requirements.txt
aiohttp
aiortc
opencv-python
gdown
dialogflow
hydra-core
hydra-colorlog
numpy
scipy
pydub
uvicorn
fastapi
starlette
